refactor(a2a-t-prompt): add SectionedTemplateRenderer SPI with collapse and drop policies
